// front/src/providers/api-client.generated/model/newsletterDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CandidateDepartmentDto } from './candidateDepartmentDto';
import { NewsLetterCandidateJobsDto } from './newsLetterCandidateJobsDto';
import { NewsletterJobOfferDto } from './newsletterJobOfferDto';
import { NewsLetterCandidateStatusDto } from './newsLetterCandidateStatusDto';
import { AppValueDto } from './appValueDto';


export interface NewsletterDto { 
    id?: string;
    creationDate?: Date;
    modifDate?: Date;
    disabled?: boolean;
    title?: string;
    content?: string;
    subject?: string;
    sender?: string;
    newsletterStatusId?: string;
    newsletterStatus?: AppValueDto;
    sendDate?: Date;
    candidatesCount?: number;
    candidateApplicationsCount?: number;
    newsLettersCandidateStatus?: Array<NewsLetterCandidateStatusDto>;
    newsLettersJob?: Array<NewsLetterCandidateJobsDto>;
    newslettersJobOffer?: Array<NewsletterJobOfferDto>;
    language?: string;
    newsletterSibId?: string;
    type?: string;
    newsletterListSibId?: string;
    sentCount?: number;
    deliveredCount?: number;
    answeredCount?: number;
    unsubscriptionsCount?: number;
    openedCount?: number;
    clickedCount?: number;
    htmlFullContent?: string;
    includeCandidateApplications?: boolean;
    candidateDepartments?: Array<CandidateDepartmentDto>;
    cityFilter?: Array<string>;
    countriesFilter?: string;
}